in work

+++ firmware/esp32/main/Catalog.cpp
@@ -4,7 +4,7 @@
 #include <sdkconfig.h>
 // override LOG_LOCAL_LEVEL per project setting
 #undef LOG_LOCAL_LEVEL
-#define LOG_LOCAL_LEVEL     CONFIG_SNEAKERNET_LOG_LEVEL
+#define LOG_LOCAL_LEVEL CONFIG_SNEAKERNET_LOG_LEVEL
 
 #include <utime.h>
 #include <cstring>
@@ -13,7 +13,7 @@ Catalog::Catalog(const std::filesystem::path _root)
     : root(_root)
 {
     // cleanup catalog state
-    for(const auto& entry : std::filesystem::recursive_directory_iterator(root))
+    for (const auto &entry : std::filesystem::recursive_directory_iterator(root))
     {
         // track the latest filetime
         if (entry.last_write_time() > latest_timestamp)
@@ -28,10 +28,7 @@ Catalog::Catalog(const std::filesystem::path _root)
             std::filesystem::remove(entry, ec);
             if (ec)
             {
-                ESP_LOGW(TAG, "failed to remove %s [%s]"
-                     ,entry.path().c_str()
-                     ,ec.message().c_str()
-                );
+                ESP_LOGW(TAG, "failed to remove %s [%s]", entry.path().c_str(), ec.message().c_str());
             }
         }
         // TODO remove any TITLE files without a file
@@ -39,16 +36,16 @@ Catalog::Catalog(const std::filesystem::path _root)
     }
 }
 
-bool Catalog::hasFolder(const std::filesystem::path& folderpath) const
+bool Catalog::hasFolder(const std::filesystem::path &folderpath) const
 {
     if (isHidden(folderpath))
         return false;
 
     std::error_code ec;
-    return std::filesystem::is_directory(root/folderpath, ec);
+    return std::filesystem::is_directory(root / folderpath, ec);
 }
 
-bool Catalog::isLocked(const std::filesystem::path& folderpath) const
+bool Catalog::isLocked(const std::filesystem::path &folderpath) const
 {
     if (isHidden(folderpath))
         return true;
@@ -56,22 +53,22 @@ bool Catalog::isLocked(const std::filesystem::path& folderpath) const
     // TODO assert that parents are not locked
 
     std::error_code ec;
-    return std::filesystem::is_regular_file(root/folderpath/LOCKED_FILENAME, ec);
+    return std::filesystem::is_regular_file(root / folderpath / LOCKED_FILENAME, ec);
 }
 
-Catalog::FolderInfo Catalog::folderInfo(const std::filesystem::path& folderpath) const
+Catalog::FolderInfo Catalog::folderInfo(const std::filesystem::path &folderpath) const
 {
     FolderInfo ret;
 
     if (isHidden(folderpath))
         return ret;
 
-    for (auto& entry : std::filesystem::directory_iterator(root/folderpath))
+    for (auto &entry : std::filesystem::directory_iterator(root / folderpath))
     {
         if (isHidden(entry.path()))
             continue;
 
-        if (entry.is_directory()) 
+        if (entry.is_directory())
             ret.subfolders.push_back(entry.path().filename());
 
         if (entry.is_regular_file())
@@ -80,8 +77,8 @@ Catalog::FolderInfo Catalog::folderInfo(const std::filesystem::path& folderpath)
             fileInfo.name = entry.path().filename();
             fileInfo.size = entry.file_size();
             fileInfo.timestamp = entry.last_write_time();
-            fileInfo.title = getTitle(folderpath/fileInfo.name);
-            fileInfo.hasIcon = hasIcon(folderpath/fileInfo.name);
+            fileInfo.title = getTitle(folderpath / fileInfo.name);
+            fileInfo.hasIcon = hasIcon(folderpath / fileInfo.name);
 
             ret.files.push_back(fileInfo);
         }
@@ -90,16 +87,41 @@ Catalog::FolderInfo Catalog::folderInfo(const std::filesystem::path& folderpath)
     return ret;
 }
 
-bool Catalog::removeFolder(const std::filesystem::path& folderpath)
+bool Catalog::addFolder(const std::filesystem::path &folderpath)
 {
+    // don't allow catalog folders to conflict with hidden files/folders
     if (isHidden(folderpath))
         return false;
 
     std::error_code ec;
-    return std::filesystem::remove(root/folderpath, ec);
+    bool ret = std::filesystem::create_directory(root / folderpath, ec);
+    if (!ret)
+        ESP_LOGE(TAG, "failed to create directory [%s] (error: %s)",
+                 folderpath.c_str(), ec.message().c_str());
+
+    return ret;
 }
 
-bool Catalog::hasFile(const std::filesystem::path& filepath) const
+bool Catalog::removeFolder(const std::filesystem::path &folderpath)
+{
+    // never remove the root folder
+    if (folderpath.empty())
+        return false;
+
+    // don't allow catalog folders to conflict with hidden files/folders
+    if (isHidden(folderpath))
+        return false;
+
+    std::error_code ec;
+    bool ret = std::filesystem::remove(root / folderpath, ec);
+    if (!ret)
+        ESP_LOGE(TAG, "failed to remove directory [%s] (error: %s)",
+                 folderpath.c_str(), ec.message().c_str());
+
+    return ret;
+}
+
+bool Catalog::hasFile(const std::filesystem::path &filepath) const
 {
     if (isHidden(filepath))
     {
@@ -108,31 +130,31 @@ bool Catalog::hasFile(const std::filesystem::path& filepath) const
     }
 
     std::error_code ec;
-    auto ret = std::filesystem::is_regular_file(root/filepath, ec);
+    auto ret = std::filesystem::is_regular_file(root / filepath, ec);
     if (ec)
         ESP_LOGD(TAG, "failed to determine if hasFile [%s, ec: %s]", filepath.c_str(), ec.message().c_str());
     return ret;
 }
 
-std::filesystem::file_time_type Catalog::timestamp(const std::filesystem::path& filepath) const
+std::filesystem::file_time_type Catalog::timestamp(const std::filesystem::path &filepath) const
 {
     std::error_code ec;
-    auto ret = std::filesystem::last_write_time(root/filepath, ec);
+    auto ret = std::filesystem::last_write_time(root / filepath, ec);
     if (ec)
         ESP_LOGW(TAG, "unable to get last file write time [%s]", ec.message().c_str());
     return ret;
 }
 
-std::ifstream Catalog::readContent(const std::filesystem::path& filepath) const
+std::ifstream Catalog::readContent(const std::filesystem::path &filepath) const
 {
     if (isHidden(filepath))
         // return a null stream
         return std::ifstream();
 
-    return std::ifstream(root/filepath, std::ios_base::in | std::ios_base::binary);
+    return std::ifstream(root / filepath, std::ios_base::in | std::ios_base::binary);
 }
 
-std::ifstream Catalog::readIcon(const std::filesystem::path& filepath) const
+std::ifstream Catalog::readIcon(const std::filesystem::path &filepath) const
 {
     if (isHidden(filepath))
         // return a null stream
@@ -148,16 +170,16 @@ std::ifstream Catalog::readIcon(const std::filesystem::path& filepath) const
     return std::ifstream();
 }
 
-bool Catalog::removeFile(const std::filesystem::path& filepath) const
+bool Catalog::removeFile(const std::filesystem::path &filepath) const
 {
     if (isHidden(filepath))
         return false;
 
     std::error_code ec;
-    if (std::filesystem::is_regular_file(root/filepath, ec))
+    if (std::filesystem::is_regular_file(root / filepath, ec))
     {
         // remove the file
-        std::filesystem::remove(root/filepath, ec);
+        std::filesystem::remove(root / filepath, ec);
         if (ec)
             ESP_LOGW(TAG, "failed to remove file [%s ec:%s]", filepath.c_str(), ec.message().c_str());
 
@@ -175,14 +197,13 @@ bool Catalog::removeFile(const std::filesystem::path& filepath) const
 
 // Upload Support
 //==============================================================================
-Catalog::InWorkContent::InWorkContent(const std::filesystem::path& filepath, const std::optional<std::filesystem::file_time_type> timestamp)
-    : filepath(filepath)
-     ,timestamp(timestamp)
+Catalog::InWorkContent::InWorkContent(const std::filesystem::path &filepath, const std::optional<std::filesystem::file_time_type> timestamp)
+    : filepath(filepath), timestamp(timestamp)
 {
     auto parentpath = filepath.parent_path();
     auto filename = filepath.filename().string();
     auto inwork_filename = filename.insert(0, INWORK_PREFIX);
-    inwork_filepath = parentpath/inwork_filename;
+    inwork_filepath = parentpath / inwork_filename;
     ofs = std::ofstream(inwork_filepath, std::ios_base::out | std::ios_base::binary);
 }
 
@@ -203,24 +224,23 @@ void Catalog::InWorkContent::done()
     std::filesystem::rename(inwork_filepath, filepath, ec);
     if (ec)
         ESP_LOGW(TAG, "unable to rename inwork file [from:%s to:%s ec:%s]",
-            inwork_filepath.c_str(),
-            filepath.c_str(),
-            ec.message().c_str()
-        );
+                 inwork_filepath.c_str(),
+                 filepath.c_str(),
+                 ec.message().c_str());
 
     if (timestamp)
     {
         // set the timestamp
         std::filesystem::last_write_time(filepath, timestamp.value(), ec);
-        if (ec) {
+        if (ec)
+        {
             // FIXME
             ESP_LOGD(TAG, "unable to set timestamp [%s ec:%s], will attemp via C api", filepath.c_str(), ec.message().c_str());
 
             // use c api (since not implemented in std::filesystem)
             const auto timestamp_s = std::chrono::system_clock::to_time_t(
-                std::chrono::file_clock::to_sys(timestamp.value())
-            );
-            struct utimbuf c_timestamp { .actime = timestamp_s, .modtime = timestamp_s };
+                std::chrono::file_clock::to_sys(timestamp.value()));
+            struct utimbuf c_timestamp{.actime = timestamp_s, .modtime = timestamp_s};
             if (0 != utime(filepath.c_str(), &c_timestamp))
             {
                 ESP_LOGW(TAG, "unable to set timestamp [%s ec:%s]", filepath.c_str(), std::strerror(errno));
@@ -242,11 +262,11 @@ Catalog::InWorkContent::~InWorkContent()
     }
 }
 
-Catalog::InWorkContent Catalog::addFile(const std::filesystem::path& filepath, const std::optional<std::filesystem::file_time_type> timestamp)
+Catalog::InWorkContent Catalog::addFile(const std::filesystem::path &filepath, const std::optional<std::filesystem::file_time_type> timestamp, const size_t size)
 {
     // create folders if not present
     std::error_code ec;
-    std::filesystem::create_directories(root/(filepath.parent_path()), ec);
+    std::filesystem::create_directories(root / (filepath.parent_path()), ec);
     if (ec)
         ESP_LOGW(TAG, "create directories failed [%s]", ec.message().c_str());
 
@@ -254,36 +274,34 @@ Catalog::InWorkContent Catalog::addFile(const std::filesystem::path& filepath, c
     {
         if (timestamp > latest_timestamp)
             latest_timestamp = timestamp.value();
-        return InWorkContent(root/filepath, timestamp);
+        return InWorkContent(root / filepath, timestamp);
     }
     else
-        return InWorkContent(root/filepath, latest_timestamp);
+        return InWorkContent(root / filepath, latest_timestamp);
 }
 
-Catalog::InWorkContent Catalog::addIcon(const std::filesystem::path& filepath)
+Catalog::InWorkContent Catalog::addIcon(const std::filesystem::path &filepath)
 {
     // FIXME use generic filepath->icon
     auto iconpath = ICON_PREFIX;
-    return InWorkContent(root/iconpath, std::nullopt);
+    return InWorkContent(root / iconpath, std::nullopt);
 }
-
 
 // PRIVATE
 //==============================================================================
 
-bool Catalog::isHidden(const std::filesystem::path& path)
+bool Catalog::isHidden(const std::filesystem::path &path)
 {
-    for(const auto& i : path)
+    for (const auto &i : path)
     {
         // ignore hidden files/folders and relative paths
         if (*(i.c_str()) == HIDDEN_PREFIX)
             return true;
-
     }
     return false;
 }
 
-std::optional<std::string> Catalog::getTitle(const std::filesystem::path& filepath) const
+std::optional<std::string> Catalog::getTitle(const std::filesystem::path &filepath) const
 {
     if (isHidden(filepath))
         return std::nullopt;
@@ -292,14 +310,14 @@ std::optional<std::string> Catalog::getTitle(const std::filesystem::path& filepa
     auto parentpath = std::filesystem::path(filepath).parent_path();
     auto filename = std::filesystem::path(filepath).filename().string();
     auto titlefilename = filename.insert(0, TITLE_PREFIX);
-    auto titlepath = root/parentpath/titlefilename;
+    auto titlepath = root / parentpath / titlefilename;
     ESP_LOGI(TAG, "looking for title in %s", titlepath.c_str());
 
     std::error_code ec;
     if (std::filesystem::is_regular_file(titlepath, ec))
     {
         std::ifstream ifs(titlepath);
-        std::stringstream ss; 
+        std::stringstream ss;
         ss << ifs.rdbuf();
         ESP_LOGI(TAG, "found title [%s]", ss.str().c_str());
         return ss.str();
@@ -309,7 +327,7 @@ std::optional<std::string> Catalog::getTitle(const std::filesystem::path& filepa
     return std::nullopt;
 }
 
-bool Catalog::hasIcon(const std::filesystem::path& filepath) const
+bool Catalog::hasIcon(const std::filesystem::path &filepath) const
 {
     if (isHidden(filepath))
         return false;
@@ -326,4 +344,3 @@ bool Catalog::hasIcon(const std::filesystem::path& filepath) const
     // std::error_code ec;
     // return std::filesystem::is_regular_file(iconpath, ec);
 }
-
